Don't accidentally mutate the light/dark theme when creating a new one (#3078)
It turns out that we were accidentally mutating the "starterTheme"'s
`emotion.inSidebar` field since the spread operator only does a shallow
copy, and `lodash/merge` mutates the target object. This resulted in a
spacing bug where, when a custom theme is set, the base theme for it
(which depends on the background color of the custom theme) has weird
spacing since it has `emotion.inSidebar` set to true for it, even in the
main content area.