Give people (mostly) minimal theme toml definitions (#3103)
Now that we can define themes by specifying a base theme (or having the
base theme default to "light") and selectively overwriting options,
there are often multiple ways to define the same theme. This change
improves the "Copy theme to clipboard" button to give the user a smaller
toml snippet defining their theme when possible.

The toml snippets now:
1. Don't specify a font unless it's changed from "sans serif"
2. Set a "light" or "dark" base (depending on what the currently defined
   custom theme is closer to) and partially define the changed
   properties when doing so produces a minimal config.
3. May not set a base at all if enough colors are changed so that doing
   so is not needed.

The reason we give people "mostly" minimal theme toml definitions is
that we include `base="light"` even when it's not strictly required to
be more explicit.